# Faster epipolar metrics