A while ago I told @edgarriba and @ducha-aiki that there are some inefficiencies in the epipolar metrics code. This PR proposes a fix, replacing inefficient "handcrafted" implementations of norms with matmul (specifically, `torch.einsum`). A [benchmark on colab](https://colab.research.google.com/drive/1RqXQIq-wL4GRcvJHNTNjAi3ckxEMgKi6?usp=sharing) shows my implementation to be ~100% faster on CPU and ~60% faster on CUDA.

While the PR is ready to be merged as is, there are more such inefficiencies which we may want to address. Aside from similar improvements to `sampson_epipolar_distance`, I found the following generic functions in `linalg` which use the same inefficient implementations: [`batched_dot_product`](https://github.com/kornia/kornia/blob/master/kornia/geometry/linalg.py#L232), [`batched_squared_norm`](https://github.com/kornia/kornia/blob/master/kornia/geometry/linalg.py#L239) and [`euclidean_distance`](https://github.com/kornia/kornia/blob/master/kornia/geometry/linalg.py#L244). It probably makes most sense to reimplement the epipolar metrics in terms of those `linalg` helpers and fix the helpers themselves. The problem with fixing them though is that I am not 100% sure my `einsum`-based alternatives will have the exact same broadcasting semantics across potential input shapes. Can we assume the test suite is extensive enough to catch these cases?

Fixes (partially) #2227 


#### Type of change
- [x] 🐞 Bug fix (non-breaking change which fixes an issue)


#### Checklist

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] My changes generate no new warnings