@edgarriba 2023-05-10
For the rest of your proposal , I would say just give it a shot, in principle we enough test cases to discover any inconsistencies. Besides, do you have any intuition why einsum is faster by default? I was told that Internally is using the same exact operators. Plus, I think is work to compare using torch.compile sine it’s the new way to make things really fast.