@jatentaki 2023-05-12
> Why is it faster?
The old code is achieving matmuls: batching aside, `(a * b).sum()` is equivalent to `torch.dot(a, b)`. This batching is sometimes tricky/unreadable to get with `torch.matmul` so I use `einsum`. The speed gains are therefore from using proper matmul kernels vs "implementing" it manually. I am not sure if `compile` is smart enough to do this kind of optimisation. In any case the biggest savings should actually be in terms of memory not speed, since we avoid keeping the intermediate (pre-reduction) matrix in memory for backwards (when relevant).

> Give it a shot
I will over the coming couple of days.