@gau-nernst 2023-05-13
Related issue from pytorch core: https://github.com/pytorch/pytorch/issues/18027

I think `torch.einsum()` might not always be faster than `(a * b).sum(-1)`? (Perhaps also depends on the vector size and PyTorch version). I think the speed up is more because `.norm().pow(2)` is replaced with dot product.

From PyTorch 1.13, there is built-in batched dot product op [`torch.linalg.vecdot()`](https://pytorch.org/docs/1.13/generated/torch.linalg.vecdot.html)