@edgarriba 2023-05-16
as @gau-nernst suggest i'd advocate to find alternatives to einsum and see how does it scale e.g later with torch compile, or onnx