@jatentaki 2023-05-16
Is vectordot good in your eyes? What is the problem with einsum exactly?