@gau-nernst 2023-05-16
@jatentaki I looked at your Google Colab. Timing with tqdm is probably not really reliable. Also, for CUDA code, `torch.cuda.synchronize()` is required. I edited your notebooks and obtained the following

Device | Old | New
---|---|---
CPU | 30 it/s | 37 it/s
CPU (compile) | 184 it/s | 55 it/s
GPU | 468 it/s | 264 it/s
GPU (compile) | 1014 it/s | 254 it/s

Hardware is from Google Colab. It doesn't seem to be faster on CUDA. And einsum seems to be not friendly with torch.compile. I also tried `torch.linalg.vecdot()` but it's slower, probably because it is optimised for larger vectors. We probably need to investigate it more.

Anyway, I think careful benchmark is needed to justify a new implementation.

The code I used

```python
import time

N = 100

# warmup
_dists = K.geometry.epipolar.symmetrical_epipolar_distance(kp1, kp2, F)

time0 = time.perf_counter()
for _ in range(N):
  _dists = K.geometry.epipolar.symmetrical_epipolar_distance(kp1, kp2, F)
print(N / (time.perf_counter() - time0))

f = torch.compile(K.geometry.epipolar.symmetrical_epipolar_distance)
_dists = f(kp1, kp2, F)

time0 = time.perf_counter()
for _ in range(N):
  _dists = f(kp1, kp2, F)
print(N / (time.perf_counter() - time0))

# warmup
_dists = faster_symmetrical_epipolar_distance(kp1, kp2, F)

time0 = time.perf_counter()
for _ in range(N):
  _dists = faster_symmetrical_epipolar_distance(kp1, kp2, F)
print(N / (time.perf_counter() - time0))

f = torch.compile(faster_symmetrical_epipolar_distance)
_dists = f(kp1, kp2, F)

time0 = time.perf_counter()
for _ in range(N):
  _dists = f(kp1, kp2, F)
print(N / (time.perf_counter() - time0))

kp1 = kp1.cuda()
kp2 = kp2.cuda()
F = F.cuda()

# warmup
_dists = K.geometry.epipolar.symmetrical_epipolar_distance(kp1, kp2, F)
torch.cuda.synchronize()

time0 = time.perf_counter()
for _ in range(N):
  _dists = K.geometry.epipolar.symmetrical_epipolar_distance(kp1, kp2, F)
  torch.cuda.synchronize()

print(N / (time.perf_counter() - time0))

f = torch.compile(K.geometry.epipolar.symmetrical_epipolar_distance)
_dists = f(kp1, kp2, F)
torch.cuda.synchronize()

time0 = time.perf_counter()
for _ in range(N):
  _dists = f(kp1, kp2, F)
  torch.cuda.synchronize()

print(N / (time.perf_counter() - time0))
```